rename app context file
